Add files via upload
I have hereby attached the report in both doc and ppt format along with the reccomendations. Thus I have also attached the dataset that I have used and the .py file that can be used to get the desired results and visualisations that are displayed in the reports. I have made sure to adhere to the guidelines provided The .py file can be accessed through Google Colab for optimum visualisations